+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 - Timed choice
 - Image tag
+- Target a character using @char(name) in the command tag (e.g. `# command "attribute @char(user) minecraft:scale base set 3"`)
 
 ### Fixes
 - Dialog 2D scale not fixed based on gui scale